dnd cards around done
