+++ productab_client/src/redux/reducers/cardReducer.js
@@ -1,28 +1,5 @@
 import * as _ from "../../constants";
-
-const reorder = (cards, startIndex, endIndex) => {
-    const result = Array.from(cards);
-    const [removed] = result.splice(startIndex, 1);
-    result.splice(endIndex, 0, removed);
-    return result;
-};
-
-// Moves an item from one list to another list.
-const move = (source, destination, cards) => {
-  // if (source.droppableId === destination.droppableId) {
-  //   const result = Array.from(cards);
-  //   const [removed] = result.splice(source.index, 1);
-  //   result.splice(destination.index, 0, removed);
-  //   return result;
-  // }
-    const sourceClone = cards.filter(card => card.list_id === source.droppableId);
-    const destClone = cards.filter(card => card.list_id === destination.droppableId);
-    const [removed] = sourceClone.splice(source.index, 1);
-    removed.list_id = destination.droppableId
-    destClone.splice(destination.index, 0, removed);
-    let results = [...sourceClone, ...destClone]
-    return results;
-};
+import { moveCards } from "./moveElements";
 
 const initialState = {
   selectedCard: null,
@@ -49,13 +26,12 @@ export const card = (state = initialState, {type, payload}) => {
 
     case _.MOVE_CARD:
       if (payload.source && payload.destination) {
-        const cards = move(
+        const cards = moveCards(
           payload.source,
           payload.destination,
           state.cards
         );
-      console.log("cards", cards)
-      return { ...state, ...cards };
+      return { ...state, cards };
       }
 
     default:
